Positionen werden jetzt zum Server geschickt (Erster Versuch.. Genauigkeit? Zu oft?)

+++ src/main/java/communication/PositionUpdate.java
@@ -3,15 +3,15 @@ package communication;
 /**
  * Beschreibt das Update einer Position
  *
- * @author simon, 10.04.2016.
+ * @author Simon Danner, 10.04.2016.
  */
 public class PositionUpdate
 {
   private String phoneNumber;
-  private long longitude;
-  private long latitude;
+  private double longitude;
+  private double latitude;
 
-  public PositionUpdate(String pPhoneNumber, long pLongitude, long pLatitude)
+  public PositionUpdate(String pPhoneNumber, double pLongitude, double pLatitude)
   {
     phoneNumber = pPhoneNumber;
     longitude = pLongitude;
